# api/openapi/spec.yaml
openapi: 3.0.3
info:
  title: Bookit API
  description: |
    Bookit is a multi-vertical booking platform for beauty, sport, and pet care services.

    ## Authentication

    Most endpoints require authentication via JWT Bearer token.

    1. Register or login to get access and refresh tokens
    2. Include access token in Authorization header: `Bearer <access_token>`
    3. When access token expires, use refresh endpoint to get new tokens

    ## Rate Limiting

    API requests are rate limited. Current limits:
    - Anonymous: 100 requests/minute
    - Authenticated: 1000 requests/minute
  version: 1.0.0
  contact:
    name: Bookit Support
    email: support@bookit.app
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api-staging.bookit.app
    description: Staging
  - url: https://api.bookit.app
    description: Production
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: Create a new user account with email and password. A verification email will be sent to confirm the email address.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully. The user object will have email_verified set to false until verified via email link.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Exchange a valid refresh token for new access and refresh tokens
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout and revoke refresh token
      operationId: logout
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: Logged out successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email address
      description: Verify user's email using the token sent via email
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/resend-verification:
    post:
      tags: [Auth]
      summary: Resend verification email
      description: Request a new verification email. Requires authentication.
      operationId: resendVerificationEmail
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Verification email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Email already verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: Too many requests - rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset
      description: Send a password reset email to the user
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset email sent (always returns success to prevent email enumeration)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token
      description: Reset user's password using the token from email
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/app-switch-token:
    post:
      tags: [Auth]
      summary: Create app switch token
      description: Generate a one-time token for switching between consumer and business apps while preserving the session. Token expires in 5 minutes.
      operationId: createAppSwitchToken
      security:
        - BearerAuth: []
      responses:
        '200':
          description: App switch token created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSwitchTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/exchange-app-switch-token:
    post:
      tags: [Auth]
      summary: Exchange app switch token for session
      description: Exchange a one-time app switch token for a full authentication session. The token is invalidated after use.
      operationId: exchangeAppSwitchToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeAppSwitchTokenRequest'
      responses:
        '200':
          description: Token exchanged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================================
  # USERS
  # ============================================================
...
  /api/v1/providers:
    post:
      tags: [Providers]
      summary: Become a provider
      description: Upgrade current user account to provider status
      operationId: createProvider
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Provider account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: User is already a provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/providers/me:
    get:
      tags: [Providers]
      summary: Get current provider profile
      operationId: getCurrentProvider
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Provider profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: User is not a provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================================
  # BUSINESSES
  # ============================================================
  /api/v1/businesses:
    get:
      tags: [Businesses]
      summary: List my businesses
      description: List all businesses owned by the current provider
      operationId: listBusinesses
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: List of businesses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ProviderRequired'

    post:
      tags: [Businesses]
      summary: Create a business
      operationId: createBusiness
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessCreate'
      responses:
        '201':
          description: Business created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ProviderRequired'
  /api/v1/businesses/{id}:
    get:
      tags: [Businesses]
      summary: Get a business
      operationId: getBusiness
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BusinessId'
      responses:
        '200':
          description: Business details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ProviderRequired'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Businesses]
      summary: Update a business
      operationId: updateBusiness
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BusinessId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessUpdate'
      responses:
        '200':
          description: Business updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ProviderRequired'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Businesses]
      summary: Delete a business
      operationId: deleteBusiness
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BusinessId'
      responses:
        '204':
          description: Business deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ProviderRequired'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/businesses/{id}/logo:
    post:
      tags: [Businesses]
      summary: Upload business logo
      operationId: uploadBusinessLogo
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BusinessId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Logo uploaded, returns updated business
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ProviderRequired'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # LOCATIONS
  # ============================================================
  /api/v1/locations:
    get:
      tags: [Locations]
      summary: List locations
      description: List locations for a business
      operationId: listLocations
      security:
        - BearerAuth: []
      parameters:
        - name: business_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: List of locations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ProviderRequired'

    post:
      tags: [Locations]
      summary: Create a location
      operationId: createLocation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationCreate'
      responses:
        '201':
          description: Location created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ProviderRequired'
  /api/v1/locations/{id}:
    get:
      tags: [Locations]
      summary: Get a location
      operationId: getLocation
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LocationId'
      responses:
        '200':
          description: Location details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Locations]
      summary: Update a location
      operationId: updateLocation
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LocationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationUpdate'
      responses:
        '200':
          description: Location updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ProviderRequired'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Locations]
      summary: Delete a location
      operationId: deleteLocation
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LocationId'
      responses:
        '204':
          description: Location deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ProviderRequired'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/locations/{id}/schedule:
    get:
      tags: [Locations]
      summary: Get location schedule
      operationId: getLocationSchedule
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LocationId'
      responses:
        '200':
          description: Location schedule with days and upcoming exceptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
        - phone
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          example: "securepassword123"
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "John Doe"
        phone:
          type: string
          example: "+37061234567"
...
    AuthResponse:
      type: object
      required:
        - user
        - tokens
      properties:
        user:
          $ref: '#/components/schemas/User'
        tokens:
          $ref: '#/components/schemas/AuthTokens'
...